More work on getting SceneStateSelectMode matching
