+++ lib/rs/KidsMode.h
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace rs {
+bool isKidsMode(const al::LayoutActor*);
+};
