+++ src/al/input/CMakeLists.txt
@@ -1,4 +1,3 @@
 target_sources(odyssey PRIVATE
-    PlayerInput.cpp
-    PlayerInputFunction.cpp
+    KeyRepeatCtrl.cpp
 )
