+++ src/game/player/CMakeLists.txt
@@ -1,3 +1,5 @@
+add_subdirectory(input)
+
 target_sources(odyssey PRIVATE
     PlayerEquipmentUser.cpp
     PlayerInfo.cpp
