+++ src/game/player/input/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(odyssey PRIVATE
+    PlayerInput.cpp
+    PlayerInputFunction.cpp
+)
